Major debugging - code now working
Rewritten function for calculating subsurface reflections - now using while logic instead of division by beam_d_wat because I realised the previous logic was flawed (the depth gained is not the same in the first and subsequent reflections because the first wall strike occurs after the beam has entered the water at the SurfStrike_d, whereas all subsequent strikes occur after the beam has reflected from the wall and traversed a horizontal distance equal to the hold width). Now behaves sensibly in unit tests and when run from the driver.