Move functions to external class, include absorption losses in water
Main update is addition of losses dueto extinction as beam passes through water. The total path length for the beam is calculated taking into account all reflectionsin the water column - this is multiplied by the absorption coefficient to give optical thickness which is then used to attenuate beam. Also, functions moved to external files and generall tidying and annotating.